Add support for the metric system when selecting the thickness of lines for elements on the stage

+++ plugins/robots/common/twoDModel/src/engine/items/stylusItem.cpp
@@ -115,7 +115,7 @@ void StylusItem::setPenStyle(const QString& text)
 	mStylusImpl.setPenStyle(mAbstractListLine, text);
 }
 
-void StylusItem::setPenWidth(int width)
+void StylusItem::setPenWidth(qreal width)
 {
 	AbstractItem::setPenWidth(width);
 	mStylusImpl.setPenWidth(mAbstractListLine, width);
